feat : 특정 시간에 대한 예약이 존재할 때 삭제 시 예외 발생 기능 추가

+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -36,6 +36,10 @@ public class ReservationTimeService {
     }
 
     public void delete(Long id) {
+        findById(id);
+        if (reservationTimeRepository.findReservationInSameId(id).isPresent()) {
+            throw new IllegalArgumentException("이미 해당 시간에 예약이 있습니다.");
+        }
         reservationTimeRepository.delete(id);
     }
 }
